replace sandbox with localnet
