+++ beaker/sandbox/__init__.py
@@ -1,5 +1,6 @@
 from .clients import get_algod_client, get_indexer_client
 from .kmd import SandboxAccount, add_account, get_accounts
+from warnings import warn
 
 __all__ = [
     "SandboxAccount",
@@ -8,3 +9,5 @@ __all__ = [
     "get_algod_client",
     "get_indexer_client",
 ]
+
+warn("beaker.sandbox is being deprecated, use beaker.localnet instead")
