+++ examples/opup/demo.py
@@ -7,7 +7,7 @@ from algosdk.atomic_transaction_composer import (
 
 from beaker.client import ApplicationClient
 from beaker.consts import milli_algo
-from beaker.sandbox import get_accounts, get_algod_client
+from beaker.localnet import get_accounts, get_algod_client
 
 from examples.opup import contract
